added failfast

+++ tests/test_executable.py
@@ -12,10 +12,11 @@ class PstTestCase(unittest.TestCase):
     def tearDownClass(cls):
         files = ["trees-root.txt", "trees-pid.txt", "trees-pst.txt"]
         for f in files:
-            try:
-                os.remove(f)
-            except BaseException:
-                print("Error while deleting file ", f)
+	    if os.path.isfile(f):
+                try:
+                    os.remove(f)
+                except BaseException:
+                    print("Error while deleting file ", f)
 
     def test_file_exists(self):
         self.assertTrue(find_executable("pst"))
@@ -24,7 +25,7 @@ class PstTestCase(unittest.TestCase):
         self.assertEqual(os.access(find_executable("pst"), os.X_OK), True)
 
     def test_compare_file_size(self):
-        self.assertEqual(os.path.getsize(find_executable("pst")), os.path.getsize('bin/pst'))
+        self.assertEqual(os.path.getsize(find_executable("pst")), os.path.getsize('../bin/pst'))
 
     def test_help_string(self):
         proc = Popen(["pst", "-h"], stdout=PIPE, stderr=PIPE)
@@ -54,4 +55,4 @@ class PstTestCase(unittest.TestCase):
         self.assertTrue(os.path.getsize('trees-pst.txt') > os.path.getsize('trees-pid.txt')) 
 
 if __name__ == "__main__":
-    unittest.main()
+    unittest.main(failfast=True)
